remove shared query?
