+++ scripts/networks.js
@@ -1,6 +1,5 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 const path = require('path');
-const { BACKEND_NETWORKS_QUERY } = require('../src/resources/metadata/sharedQueries');
 
 const fs = require('fs-extra');
 
@@ -8,11 +7,94 @@ const fs = require('fs-extra');
  * Fetches data from the GraphQL API and saves it to a JSON file.
  */
 async function fetchData() {
+  const graphqlQuery = `
+    query getNetworks($device: Device!, $includeTestnets: Boolean!) {
+      networks(device: $device, includeTestnets: $includeTestnets) {
+        id
+        name
+        label
+        icons {
+          badgeURL
+        }
+        testnet
+        internal
+        opStack
+        defaultExplorer {
+          url
+          label
+          transactionURL
+          tokenURL
+        }
+        defaultRPC {
+          enabledDevices
+          url
+        }
+        gasUnits {
+          basic {
+            approval
+            swap
+            swapPermit
+            eoaTransfer
+            tokenTransfer
+          }
+          wrapped {
+            wrap
+            unwrap
+          }
+        }
+        nativeAsset {
+          address
+          name
+          symbol
+          decimals
+          iconURL
+          colors {
+            primary
+            fallback
+            shadow
+          }
+        }
+        nativeWrappedAsset {
+            address
+            name
+            symbol
+            decimals
+            iconURL
+            colors {
+            primary
+            fallback
+            shadow
+          }
+        }
+        enabledServices {
+          meteorology {
+            enabled
+          }
+          swap {
+            enabled
+          }
+          addys {
+            approvals
+            transactions
+            assets
+            positions
+          }
+          tokenSearch {
+            enabled
+          }
+          nftProxy {
+            enabled
+          }
+        }
+      }
+    }
+  `;
+
   const response = await fetch('https://metadata.p.rainbow.me/v1/graph', {
     method: 'POST',
     headers: { 'Content-Type': 'application/json' },
     body: JSON.stringify({
-      query: BACKEND_NETWORKS_QUERY,
+      query: graphqlQuery,
       variables: { device: 'APP', includeTestnets: true },
     }),
   });
